create global error handler at controller and middleware

+++ controller/productController.js
@@ -1,7 +1,8 @@
 const { Product } = require('../models/index');
 const imagekit = require('../lib/imagekit');
+const ApiError = require('../utils/apiError');
 
-const createProduct = async (req, res) => {
+const createProduct = async (req, res, next) => {
   const { name, price, stock } = req.body;
   const file = req.file;
 
@@ -10,10 +11,23 @@ const createProduct = async (req, res) => {
   const extension = split[split.length - 1];
 
   try {
-    const img = await imagekit.upload({
-      file: file.buffer,
-      fileName: `IMG=${Date.now()}.${extension}`,
-    });
+    if (file) {
+      // dapatkan extension file nya
+      const split = file.originalname.split('.');
+      const extension = split[split.length - 1];
+
+      // upload file ke imagekit
+      const uploadedImage = await imagekit.upload({
+        file: file.buffer,
+        fileName: `IMG-${Date.now()}.${extension}`,
+      });
+      img = uploadedImage.url;
+    }
+
+    // const img = await imagekit.upload({
+    //   file: file.buffer,
+    //   fileName: `IMG=${Date.now()}.${extension}`,
+    // });
     const newProduct = await Product.create({
       name,
       price,
@@ -26,14 +40,11 @@ const createProduct = async (req, res) => {
       data: { newProduct },
     });
   } catch (err) {
-    res.status(400).json({
-      status: 'failed',
-      message: err.message,
-    });
+    next(new ApiError(err.message, 400));
   }
 };
 
-const findProducts = async (req, res) => {
+const findProducts = async (req, res, next) => {
   try {
     const products = await Product.findAll();
 
@@ -42,14 +53,11 @@ const findProducts = async (req, res) => {
       data: { products },
     });
   } catch (err) {
-    res.status(400).json({
-      status: 'failed',
-      message: err.message,
-    });
+    next(new ApiError(err.message, 400));
   }
 };
 
-const findProductById = async (req, res) => {
+const findProductById = async (req, res, next) => {
   try {
     const product = await Product.findOne({
       where: {
@@ -62,14 +70,11 @@ const findProductById = async (req, res) => {
       data: { product },
     });
   } catch (err) {
-    res.status(400).json({
-      status: 'failed',
-      message: err.message,
-    });
+    next(new ApiError(err.message, 400));
   }
 };
 
-const updateProduct = async (req, res) => {
+const updateProduct = async (req, res, next) => {
   const { name, price, stock } = req.body;
 
   try {
@@ -91,18 +96,25 @@ const updateProduct = async (req, res) => {
       message: 'success updated product',
     });
   } catch (err) {
-    res.status(400).json({
-      status: 'failed',
-      message: err.message,
-    });
+    next(new ApiError(err.message, 400));
   }
 };
 
-const deleteProduct = async (req, res) => {
+const deleteProduct = async (req, res, next) => {
   const { name, price, stock } = req.body;
 
   try {
-    const product = await Product.destroy({
+    const product = await Product.findOne({
+      where: {
+        id: req.params.id,
+      },
+    });
+    
+    if (!product) {
+      return next(new ApiError('Product id tersebut gak ada', 404));
+    }
+
+    await Product.destroy({
       where: {
         id: req.params.id,
       },
@@ -113,10 +125,7 @@ const deleteProduct = async (req, res) => {
       message: 'success deleted product',
     });
   } catch (err) {
-    res.status(400).json({
-      status: 'failed',
-      message: err.message,
-    });
+    next(new ApiError(err.message, 400));
   }
 };
